Systems overhaul (also added Fog of War and Block map generation) (#2)
* basic working StateView + QuadTree for 'Fog of War' vision

* cleanup destroy logic

* add ore cracked effect instead of hp bar

* add fog of war rectangle

* add maxHp to ore on server

* break out constants and create Blocks (drop Enemies)

* cleanup Ore Phaser game object

* cleanup room state (don't sync server-only fields to the client)

* fix consts in game room

* create fogOverlay object

* remove enemy things

* rename Ore to Block (much better, more generic)

* update state keys for inventory

* break schemas out and add player view levels

* create a BlockMap system (handles generating a map and visible blocks per client)

* add PlayerVision system for players to see others

* refactor serverPlayer movement to happen in update() instead and using LERP

* add debug fields to player vision

* cleanup fogOverlay update() params

* update Player params to use shared type

* add FPS counter

* comment out load test functionality for now

* update tests

* cleanup client find calls (use getById instead)

* fix tests with better player snapshots (don't reference objects, duh)

* create setupVisionForClient func in PlayerVision system

* keep track of username on client

* remove demo "game results" in memory storage

* create auth system

* break out onJoin logic into Auth system and PlayerMovement system

* more auth system refactoring

* more auth system refactoring

* more auth system refactoring

* move onCreate back to top of room file

* cleanup class functions

* create playerInput system

* process player input helper

* better update of visible blocks/players (wait until moves are processed)

* refactor out movePlayer helper

* cleanup processPlayerInput helper (handles the looping, exposes callback)

* create playerMining system

* wireup playerMining system and adjust some function names

* refactor mining system into more specific functions (also improved speed by ONLY looking at visible blocks per player)

* create core-game function for checking AABB collisions (and move shared types into it's own file)

* remove extra awaits

* fix load test script

* add test for AABB collision

* add test for player vision

* fix bugs with player vision (thanks to tests!)

* add test for proper block vision

* refactor BlockMap to use 2d grid and simple lookups instead of quadtree (overkill for this use-case)

* add scope keywords to class variables

* game-api v0.0.2 - finally content with systems structure

* optimize attack collisions by only searching "nearbyBlocks"

* cleanup existing Objects for Phaser scene

* remove game results call

* cleanup console log

* add game-api as a dev dependency for desktop (to allow Schema types to be shared)

* cleanup scene helper scope

* create RoomSystem for Phaser scene

* cleanup unneeded Phaser imports

* create Block system

* create InputSystem

* create PlayerSystem

* create a UISystem

* update PingDisplay

* update Game Scene to use new systems (much cleaner, but still needs work)

* convert block array to map and id from number to string

* refactor out "clientSidePrediction" logic

* ensure setSimulationInterval matches patchRate in game-api (should help performance)

* break out PlayerSystem into RemotePlayerSystem (and separated CSP and Server Reconciliation logic)

* handle local player position interpolation (works across FPS - poor to excellent)

* cleanup logic and move into interpolateLocalPlayer method

* cleanup state for remotePlayerSystem (don't rely on Phaser entity state - avoid 'any' types)

* flip strictNullChecks back to true for game-api (seriously wtf why was this off??)

* lock down game scene systems (make private)

* ensure setupStateListeners is private by passing as a callback to setup roomEventListeners

* update private variables on game room

* reduce entanglement of backend systems (lock down private fields)

* ensure uiSystem is private by exposing callback for updating inventory (more explicit anyways)

* ensure mapBorder and mapBackground are removed on uiSystem.destroy()

* bump desktop to 0.0.2

* bump web to 0.0.2

* bump package and repo to 0.0.2

* cleanup fogOverlay styling

+++ apps/game-api/src/graphql/resolvers.ts
@@ -8,6 +8,7 @@ export const resolvers: Resolvers<Context> = {
       return true;
     },
     profile: async (_, __, { dataSources, user }) => {
+      if (!user) return null;
       return dataSources.profilesDb.getProfileByUserId(user.id);
     },
     userExists: async (_, { userName }, { dataSources }) => {
@@ -17,28 +18,25 @@ export const resolvers: Resolvers<Context> = {
     totalPlayers: async (_, __, { dataSources }) => {
       return dataSources.profilesDb.getTotalPlayers();
     },
-    gameResults: async (_, { roomId }, { dataSources }) => {
-      const gameResultObject = dataSources.gameResults[roomId];
-      const gameResultArray = Object.values(gameResultObject ?? {});
-
-      if (gameResultArray.length === 0) return null;
-      return gameResultArray;
-    },
+    gameResults: () => [],
   },
   Mutation: {
     createProfile: async (_, { userName }, { dataSources, user }) => {
+      if (!user) return null;
       return dataSources.profilesDb.createProfile({
         userId: user.id,
         userName,
       });
     },
     updateProfile: async (_, { userName }, { dataSources, user }) => {
+      if (!user) return null;
       return dataSources.profilesDb.updateProfile({
         userId: user.id,
         userName,
       });
     },
     deleteProfile: async (_, __, { authClient, dataSources, user }) => {
+      if (!user) return false;
       try {
         await dataSources.profilesDb.deleteProfile(user.id);
         await authClient.deleteUser(user.id);
